requirements.txt
flask
# openai
# groq
pandas
google-generativeai
python-dotenv
Flask-Session